Port /clan to 1.13.2

+++ src/main/java/the_fireplace/clans/clan/Clan.java
@@ -26,7 +26,7 @@ public class Clan implements Serializable {
 	private int homeDimension;
 	private int claimCount = 0;
 	private boolean isOpclan = false;
-	private int rent = 0;
+	private long rent = 0;
 	private int wins = 0;
 
 	private int losses = 0;
@@ -268,11 +268,11 @@ public class Clan implements Serializable {
 		return isOpclan;
 	}
 
-	public int getRent() {
+	public long getRent() {
 		return rent;
 	}
 
-	public void setRent(int rent) {
+	public void setRent(long rent) {
 		this.rent = rent;
 	}
